requirements.txt
aiogram>=3.4.1
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
openai>=1.51.0
langchain>=0.3.2
langchain-community>=0.3.1
langchain-openai>=0.2.2
tiktoken>=0.8.0
faiss-cpu>=1.8.0.post1
aiofiles>=23.2.1